Move the state helper classes to the "_" directory.
