+++ patterns/state/manipulator_state/main.dart
@@ -9,7 +9,7 @@ import 'states/creations/circle_creation_state.dart';
 import 'states/creations/rectangle_creation_state.dart';
 import 'states/creations/text_creation_state.dart';
 import 'states/free_sate.dart';
-import 'states/mixins/paint_style.dart';
+import 'states/_/paint_style.dart';
 import 'widgets/current_state.dart';
 import 'widgets/drawing_board.dart';
 import 'widgets/tool_bar.dart';
