+++ patterns/state/manipulator_state/states/_/paint_style.dart
