+++ patterns/state/manipulator_state/states/_/sub_states/parent_state.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
-import '../sub_states/child_state.dart';
-import '../selection_state.dart';
+import '../../selections/selection_state.dart';
+import 'child_state.dart';
 import '../../../shapes/shape.dart';
 
 class ParentState<T extends Shape> extends SelectionState<T> {
